Implementando gestao de pedidos do restaurante

+++ ifood/cadastro/src/main/java/github/com/marcoant007/ifood/models/Restaurante.java
@@ -15,6 +15,8 @@ import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.DynamicUpdate;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+
 
 @Entity
 @DynamicUpdate
@@ -23,7 +25,7 @@ public class Restaurante {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String proprietario;
 
@@ -41,11 +43,11 @@ public class Restaurante {
     private Date dataAtualizacao;
 
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
